# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Time Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: includeInactive
          in: query
          schema:
            type: boolean
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Employees ────────────────────────────────────────────────────────────────
...
  /reports/utilization:
    get:
      operationId: getUtilizationReport
      tags: [reports]
      summary: Employee utilization report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Utilization data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UtilizationRow"
  /reports/projects:
    get:
      operationId: getProjectReport
      tags: [reports]
      summary: Project hours report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Project report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectReportRow"
  /reports/clients:
    get:
      operationId: getClientReport
      tags: [reports]
      summary: Client hours report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Client report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientReportRow"

  # ── Dashboard ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        name:
          type: string
        code:
          type: ["string", "null"]
        active:
          type: boolean
        isBillable:
          type: boolean
        budgetHours:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - clientName
        - name
        - code
        - active
        - isBillable
        - budgetHours
        - startDate
        - endDate
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        clientId:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        active:
          type: boolean
        isBillable:
          type: boolean
        budgetHours:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        color:
          type: ["string", "null"]
        pmName:
          type: ["string", "null"]
      required:
        - clientId
        - name
    UpdateProjectBody:
      type: object
      properties:
        clientId:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        active:
          type: boolean
        isBillable:
          type: boolean
        budgetHours:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        color:
          type: ["string", "null"]
        pmName:
          type: ["string", "null"]

    # ── Employee ──────────────────────────────────────────────────────────────
...
    UtilizationRow:
      type: object
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        availableHours:
          type: number
        billableHours:
          type: number
        nonBillableHours:
          type: number
        totalBookedHours:
          type: number
        billableUtilization:
          type: number
        overallUtilization:
          type: number
      required:
        - employeeId
        - employeeName
        - availableHours
        - billableHours
        - nonBillableHours
        - totalBookedHours
        - billableUtilization
        - overallUtilization
    ProjectReportRow:
      type: object
      properties:
        projectId:
          type: integer
        projectName:
          type: string
        clientName:
          type: string
        isBillable:
          type: boolean
        totalHours:
          type: number
        billableHours:
          type: number
        nonBillableHours:
          type: number
      required:
        - projectId
        - projectName
        - clientName
        - isBillable
        - totalHours
        - billableHours
        - nonBillableHours
    ClientReportRow:
      type: object
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        totalHours:
          type: number
        billableHours:
          type: number
        nonBillableHours:
          type: number
      required:
        - clientId
        - clientName
        - totalHours
        - billableHours
        - nonBillableHours
...